Cleaned compile errors in tests after last updates

+++ UnitTestsXCT/Node_tests.mm
@@ -17,7 +17,7 @@
 {
     Node n;
     const Key* k = n.getKey();
-    Key k2(0,0);
+    Key k2((uint32_t)0,0);
     XCTAssertTrue(*k==k2);
 }
